# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Appointments Manager API Documentation",
        "description": "API documentation for the Appointments Manager application.",
        "contact": {
            "email": "hello@frankpulido.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://appointmentmanager-production-5057.up.railway.app/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/practitioner-vacations": {
            "get": {
                "tags": [
                    "PractitionerVacation"
                ],
                "summary": "Get vacations for practitioners",
                "description": "Retrieve a list of vacations for practitioners. Admins can see all practitioners' vacations, while practitioners can see only their own.",
                "operationId": "618e2a3a287246a91f99ceeaf6e056b8",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of vacations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "practitioners": {
                                            "description": "List of practitioners with their IDs and names",
                                            "type": "object",
                                            "example": {
                                                "1": "John Doe",
                                                "2": "Jane Smith"
                                            }
                                        },
                                        "vacations": {
                                            "description": "Vacations grouped by practitioner ID",
                                            "type": "object",
                                            "example": {
                                                "1": [
                                                    {
                                                        "id": 1,
                                                        "practitioner_id": 1,
                                                        "start_date": "2024-08-01",
                                                        "end_date": "2024-08-10",
                                                        "created_at": "2024-07-01T12:00:00Z",
                                                        "updated_at": "2024-07-01T12:00:00Z"
                                                    }
                                                ],
                                                "2": [
                                                    {
                                                        "id": 2,
                                                        "practitioner_id": 2,
                                                        "start_date": "2024-09-05",
                                                        "end_date": "2024-09-15",
                                                        "created_at": "2024-07-02T12:00:00Z",
                                                        "updated_at": "2024-07-02T12:00:00Z"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "PractitionerVacation"
                ],
                "summary": "Update a vacation period for a practitioner",
                "description": "Update a vacation period for a practitioner.",
                "operationId": "3a147b62fe95840b3087ecea06d3f2c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "practitioner_id",
                                    "vacation_id",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "practitioner_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vacation_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-05"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El período de vacaciones se ha actualizado con éxito"
                                        },
                                        "vacation": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "practitioner_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-08-05"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-08-15"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-07-01T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-07-10T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vacation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vacation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to update vacation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "PractitionerVacation"
                ],
                "summary": "Create a new vacation period for a practitioner",
                "description": "Create a new vacation period for a practitioner.",
                "operationId": "7d336e1f1d5abddb6e872eff455b8692",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "practitioner_id",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "practitioner_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vacation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El período de vacaciones se ha creado con éxito"
                                        },
                                        "vacation": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "practitioner_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-08-01"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-08-10"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-07-01T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-07-01T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create vacation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "PractitionerVacation"
                ],
                "summary": "Delete a vacation period for a practitioner",
                "description": "Delete a vacation period for a practitioner.",
                "operationId": "4bf2ccb436602a1e8f6ebc688295d8c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "practitioner_id",
                                    "vacation_id"
                                ],
                                "properties": {
                                    "practitioner_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vacation_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacation deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El período de vacaciones se ha eliminado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}